+++ doc/todo.rb
@@ -1,7 +1,4 @@
 # frozen_string_literal: true
 
-# Search count details (clever lookup)
-Hausgold::User.where(confirmed: true).count
-
 # Query Scopes
 Hausgold::User.confirmed